add integration tests, schema migration script, and GitHub workflow for testing
