+++ app/config/settings.test.toml
@@ -1,1 +1,6 @@
 [test]
+DATABASE_USER = "postgres"
+DATABASE_PASSWORD = "postgres"
+DATABASE_NAME = "postgres"
+DATABASE_HOST = "127.0.0.1"
+DATABASE_PORT = "5432"
