+++ app/main.py
@@ -14,8 +14,7 @@ from api.v1.entity_tags import controller as entity_tags_controller
 from common.config import settings
 from common.constants import DEFAULT_APP_PORT
 from common.utils.logging_utils import setup_logger, LoggingFormat
-from alembic import command
-from alembic.config import Config
+from schema_migration import run_alembic_upgrade
 
 logger = setup_logger(level="DEBUG", fmt=LoggingFormat.CONSOLE)
 
@@ -25,26 +24,6 @@ app.include_router(tag_group_controller.router, prefix="/api/v1")
 app.include_router(general_controller.router, prefix="/api/v1")
 app.include_router(tags_controller.router, prefix="/api/v1")
 app.include_router(entity_tags_controller.router, prefix="/api/v1")
-
-
-def run_alembic_upgrade():
-    # inject values from dynoconf
-    # Alembic Config object
-    # alembic_cfg = context.config
-    # alembic_cfg.set_main_option("sqlalchemy.url", settings.get("DATABASE_URL"))
-
-    alembic_cfg = Config()
-    alembic_cfg.set_main_option("script_location", "./migrations")
-    alembic_cfg.set_main_option("sqlalchemy.url", settings.database_url)
-    # alembic_cfg = Config("alembic.ini")
-    try:
-        logger.info("Running Alembic upgrade")
-        command.upgrade(alembic_cfg, "head")
-        logger.info("Alembic upgrade completed successfully")
-    except Exception as e:
-        logger.error(f"Alembic upgrade failed: {e}")
-        raise
-
 
 if __name__ == "__main__":
     import uvicorn
