+++ docker-compose.yml
@@ -12,6 +12,11 @@ services:
       - '5432:5432'
     volumes: 
       - db:/var/lib/postgresql/data
+    healthcheck:
+      test: [ "CMD-SHELL", "sh -c 'pg_isready -U postgres -d postgres'" ]
+      interval: 1s
+      timeout: 3s
+      retries: 3
   metadata:
     image: metadata-svc:latest
     ports:
