+++ requirements.txt
@@ -12,3 +12,6 @@ black==24.8.0
 pylint==3.2.7
 python-dotenv==1.0.1
 tenacity==9.0.0
+pytest==8.3.3
+httpx==0.27.2
+starlette~=0.36.3
