+++ tests/conftest.py
@@ -1,0 +1,7 @@
+""" pytest configuration file"""
+import sys
+import os
+
+# Add the root of your project to the PYTHONPATH
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
